+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - default Timeline on Kiosk
 - Streamer-Interface
 - single_shot on timeline
+- synced apply default